Refactor: Workspace 컨트롤러 API 통합 테스트에 Audit 관련 테스트 추가
